requirements.txt
tensorflow
keras
pillow
numpy
flask